Revamp day-sunset-night (renamed from day-dusk-night) transitions
Transitions that go from day -> sunset -> night are changed so that
there is no stage where the animation is fixed on the sunset background.
The timings for the transitions are also changed so that the sunset ->
night transition is the length of time between sunset and nautical
twilight and the day -> sunset transition is the same length of time but
subtracted from the time for sunset. This avoids the hard-coded times
you were using before that made the transition wildly mismatched with
respect to what things looked like outside.

Also, fix some variable names that were unwittingly getting overwritten.